+++ bin/main.dart
@@ -1,0 +1,8 @@
+import 'nrfutil.dart' as nrfutil;
+
+void main(List<String> arguments) {
+  print(
+    'This command is deprecated and replaced with "flutter pub run nrfutil"',
+  );
+  nrfutil.main(arguments);
+}