+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Fixed: generating signing key file and signing validation. 
+* Added: Added the ability to use `dart run nrfutil`
+
 ## 0.0.2
 
 * Fixed ReadMe file and added comments.